move table logic to hooks
