+++ context/app/static/js/components/savedLists/SavedEntitiesTable/style.ts
@@ -1,5 +1,6 @@
 import { styled } from '@mui/material/styles';
 import Button from '@mui/material/Button';
+
 import { WhiteBackgroundIconButton } from 'js/shared-styles/buttons';
 
 const LeftMarginButton = styled(Button)(({ theme }) => ({
