+++ context/app/static/js/components/workspaces/Tables/InvitationsTable/index.ts
@@ -1,0 +1,3 @@
+import InvitationsTable from './InvitationsTable';
+
+export default InvitationsTable;
