+++ context/app/static/js/components/workspaces/Tables/WorkspacesTable/WorkspacesTable.tsx
@@ -1,10 +1,11 @@
-import React, { useEffect, useMemo, useState } from 'react';
+import React from 'react';
 import Stack from '@mui/material/Stack';
 import { WorkspaceWithUserId } from 'js/components/workspaces/types';
 import Description from 'js/shared-styles/sections/Description';
 import WorkspaceItemsTable from 'js/components/workspaces/Tables/WorkspaceItemsTable/WorkspaceItemsTable';
-import { SortField, TableFilter, TableField } from 'js/components/workspaces/Tables/WorkspaceItemsTable/types';
+import { SortField, TableField } from 'js/components/workspaces/Tables/WorkspaceItemsTable/types';
 import { StyledNumSelectedHeader } from 'js/components/workspaces/Tables/WorkspaceItemsTable/style';
+import useWorkspacesTable from 'js/components/workspaces/Tables/WorkspacesTable/hooks';
 
 const tableFields: TableField[] = [
   {
@@ -41,43 +42,7 @@ export default function WorkspacesTable({
   isLoading: boolean;
   toggleItem: (item: string) => void;
 }) {
-  const { ownCount, sharedCount } = useMemo(() => {
-    const own = workspacesList.filter((workspace) => !workspace.user_id).length;
-    return { ownCount: own, sharedCount: workspacesList.length - own };
-  }, [workspacesList]);
-
-  const [showOwn, setShowOwn] = useState(true);
-  const [showShared, setShowShared] = useState(true);
-
-  useEffect(() => {
-    setShowShared(sharedCount > 0);
-  }, [sharedCount]);
-
-  const filteredWorkspaces = useMemo(
-    () =>
-      [...workspacesList].filter((workspace) => {
-        return (showOwn && !workspace.user_id) || (showShared && !!workspace.user_id);
-      }),
-    [workspacesList, showOwn, showShared],
-  );
-
-  const filters: TableFilter[] = useMemo(
-    () => [
-      {
-        label: `Created by Me (${ownCount})`,
-        setShow: setShowOwn,
-        show: showOwn,
-        disabled: !ownCount,
-      },
-      {
-        label: `Created by Other (${sharedCount})`,
-        setShow: setShowShared,
-        show: showShared,
-        disabled: !sharedCount,
-      },
-    ],
-    [ownCount, sharedCount, showOwn, showShared],
-  );
+  const { filteredWorkspaces, filters } = useWorkspacesTable(workspacesList);
 
   return Object.keys(workspacesList).length === 0 ? (
     <Description>No workspaces created yet.</Description>
