+++ context/app/static/js/components/workspaces/Tables/WorkspacesTable/index.ts
@@ -1,0 +1,3 @@
+import WorkspacesTable from './WorkspacesTable';
+
+export default WorkspacesTable;
